requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.52
pydantic==1.10.9
python-multipart==0.0.6
pytest==7.4.0
requests==2.31.0
python-jose==3.3.0
httpx==0.24.1
